Secondbranch

+++ database/migrations/2023_10_01_102845_create_transactions_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void 
+    public function up(): void
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->unsignedBigInteger('category_id')->nullable();
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('set null');
             $table->string('title');
-            $table->integer('amount');
+            $table->decimal('amount', 10, 2);
             $table->date('date');
             $table->timestamps();
         });
